[inductor] break out TritonTemplate, TritonTemplateKernel, TritonTemplateCaller out of select_algorithm.py
Summary:
# Why

- simplify import and dependency structure
- simplify select_algorithm.py and keep semantically closer to actually select algorithm
- consolidate all triton template related things in triton_template/

# What

- pull out out select_algorithm.py
  - TritonTemplate
  - TritonTemplateCaller
  - TritonTemplateKernel
and move into triton_heuristics

- copy SymbolicGridFn into triton_templates/common.py

The copying here is done to simplify (I know) the PR and later migrate other users when they migrate into triton_templates/

- fix import sites across

# Context

https://docs.google.com/document/d/1UoTRm9fWvyUsuft_tN5hUl38Q6n61xAESboMdlspxnw/edit?tab=t.0#heading=h.xrhwr0nuqveh

Test Plan:
- ci testing

```
python3 /data/users/coconutruben/fbsource/fbcode/scripts/coconutruben/torchmm/experiment.py
```

```
V0617 19:48:23.589000 3601502 site-packages/torch/_inductor/select_algorithm.py:609] [0/0] Precompilation elapsed time: 0.00s
V0617 19:48:23.590000 3601502 site-packages/torch/_inductor/select_algorithm.py:638] [0/0] Autotuning elapsed time: 0.00s
AUTOTUNE mm(4096x3008, 3008x2048)
strides: [3008, 1], [2048, 1]
dtypes: torch.float16, torch.float16
  mm 0.0719 ms 100.0%
  triton_mm_persistent_tma_19 0.0738 ms 97.4% ACC_TYPE='tl.float32', ALLOW_TF32=False, A_ROW_MAJOR=True, BLOCK_K=64, BLOCK_M=128, BLOCK_N=256, B_ROW_MAJOR=True, EVEN_K=True, GROUP_M=8, NUM_SMS=132, TMA_EXPERIMENTAL_API=True, TMA_SIZE=128, USE_FAST_ACCUM=False, num_stages=3, num_warps=8
  triton_mm_17 0.0769 ms 93.4% ACC_TYPE='tl.float32', ALLOW_TF32=False, BLOCK_K=64, BLOCK_M=128, BLOCK_N=128, EVEN_K=True, GROUP_M=8, USE_FAST_ACCUM=False, num_stages=3, num_warps=4
```

[ghstack-poisoned]
