+++ torch/_inductor/fx_passes/b2b_gemm.py
@@ -1,5 +1,6 @@
 # mypy: allow-untyped-defs
 import functools
+import typing
 from collections import deque
 
 import torch
@@ -7,6 +8,7 @@ from torch.utils._ordered_set import OrderedSet
 from torch.utils._pytree import tree_map
 
 from ..._dynamo.utils import counters
+from ..codegen.triton_templates.template import TritonTemplate
 from ..ir import (
     ComputedBuffer,
     FixedLayout,
@@ -28,8 +30,6 @@ from ..select_algorithm import (
     autotune_select_algorithm,
     ExternKernelChoice,
     SymbolicGridFn,
-    TritonTemplate,
-    TritonTemplateCaller,
 )
 from ..utils import ceildiv
 
@@ -37,6 +37,9 @@ from ..utils import ceildiv
 B2B_GEMM_PASS = PatternMatcherPass(
     pass_name="b2b_gemm_pass",
 )
+
+if typing.TYPE_CHECKING:
+    from ..codegen.triton_templates.caller import TritonTemplateCaller
 
 
 @SymbolicGridFn
