+++ torch/_inductor/select_algorithm.py
@@ -1,25 +1,18 @@
 # mypy: allow-untyped-defs
-import contextlib
 import dataclasses
 import functools
 import inspect
-import itertools
 import json
 import logging
 import math
-import operator
 import os
 import re
 import sys
-import textwrap
 import time
 from collections.abc import Sequence
 from concurrent.futures import as_completed, ThreadPoolExecutor
-from io import StringIO
-from types import ModuleType
-from typing import Any, Callable, NamedTuple, Optional, TYPE_CHECKING, Union
+from typing import Any, Callable, Optional, TYPE_CHECKING, Union
 from typing_extensions import Self
-from unittest.mock import patch
 
 import sympy
 
@@ -27,60 +20,21 @@ import torch
 import torch._inductor.async_compile  # noqa: F401 required to warm up AsyncCompile pools
 from torch._dynamo.device_interface import get_interface_for_device
 from torch._dynamo.testing import rand_strided
-from torch._dynamo.utils import counters, dynamo_timed, identity, preserve_rng_state
+from torch._dynamo.utils import counters, dynamo_timed, preserve_rng_state
 from torch._inductor.utils import clear_on_fresh_cache
 from torch.utils._filelock import FileLock
 from torch.utils._ordered_set import OrderedSet
 
 from ..utils._sympy.functions import CeilDiv
 from . import config, ir
-from .autotune_process import (
-    TensorMeta,
-    TritonBenchmarkRequest,
-    TritonCPUBenchmarkRequest,
-    TritonGPUBenchmarkRequest,
-)
-from .codecache import code_hash, PersistentCache, PyCodeCache
-from .codegen.common import (
-    CSEVariable,
-    IndentedBuffer,
-    KernelTemplate,
-    OpOverrides,
-    WorkspaceArg,
-    WorkspaceZeroMode,
-)
-from .codegen.simd_kernel_features import SIMDKernelFeatures
+from .autotune_process import TritonGPUBenchmarkRequest
+from .codecache import code_hash, PersistentCache
 from .codegen.subgraph import SubgraphChoiceCaller
-from .codegen.triton import (
-    gen_common_triton_imports,
-    texpr,
-    TritonKernel,
-    TritonScheduling,
-)
-from .codegen.triton_utils import config_of, equal_1_arg_indices, signature_to_meta
-from .codegen.wrapper import pexpr
+from .codegen.triton_templates.caller import TritonTemplateCaller
 from .exc import CUDACompileError
 from .ir import ChoiceCaller, PrimitiveInfoType
-from .ops_handler import StoreMode
 from .runtime.benchmarking import benchmarker
-from .runtime.hints import DeviceProperties
-from .runtime.triton_compat import HAS_WARP_SPEC
-from .runtime.triton_heuristics import FixedGrid
-from .utils import (
-    ceildiv,
-    do_bench_using_profiling,
-    FakeIndentedBuffer,
-    get_dtype_size,
-    is_gpu,
-    Placeholder,
-    restore_stdout_stderr,
-    sympy_dot,
-    sympy_index_symbol,
-    sympy_product,
-    triton_type,
-    triton_type_to_torch,
-    unique,
-)
+from .utils import ceildiv, do_bench_using_profiling, is_gpu, restore_stdout_stderr
 from .virtualized import V
 
 
@@ -94,8 +48,6 @@ DEBUG = False
 
 if TYPE_CHECKING:
     import concurrent
-
-    from torch._inductor.codegen.simd import IterationRangesRoot
 
 
 class KernelNamespace:
@@ -190,940 +142,6 @@ class PartialRender:
         return self.code
 
 
-# This is used to store info needed for lowering each subgraph in triton
-# templates
-
-
-@dataclasses.dataclass()
-class SubgraphInfo:
-    body: IndentedBuffer
-    template_mask: Optional[str] = None
-    template_out: Optional[str] = None
-    compute: IndentedBuffer = dataclasses.field(default_factory=IndentedBuffer)
-    indexing_code: IndentedBuffer = dataclasses.field(default_factory=IndentedBuffer)
-    loads: IndentedBuffer = dataclasses.field(default_factory=IndentedBuffer)
-    stores: IndentedBuffer = dataclasses.field(default_factory=IndentedBuffer)
-    ops_handler: Optional[V.WrapperHandler] = None  # type: ignore[name-defined]
-
-    # only copied over if not None
-    range_trees: Optional[list["IterationRangesRoot"]] = None
-    numels = None  # type: ignore[var-annotated]
-
-    def __post_init__(self):
-        self.only_copy_if_non_none_fields = ("range_trees", "numels")
-
-    def to_dict(self):
-        return {
-            field.name: getattr(self, field.name) for field in dataclasses.fields(self)
-        }
-
-
-class ModificationWrapper(V.WrapperHandler):  # type: ignore[name-defined]
-    """Handles placeholder substitutions during subgraph processing."""
-
-    def __init__(
-        self,
-        kernel,
-        subgraph_number: int,
-        fixed_inputs: dict[str, Any],
-        mask: Optional[str],
-    ):
-        super().__init__(V.ops)
-        self.name = f"PlaceholderSubstitution_{subgraph_number}"
-        self.kernel = kernel
-        self.fixed_inputs = fixed_inputs
-        self.mask = mask
-
-    def load(self, name: str, index: sympy.Expr):
-        """Handle loading from tensor or fixed input."""
-        if name not in self.fixed_inputs:
-            index_str = self._process_indexing(index)
-            var = self._add_kernel_input(name)
-            var_dtype = V.graph.get_buffer(name).dtype
-            line = f"tl.load({var} + {index_str})"
-
-            if (
-                var_dtype in (torch.float16, torch.bfloat16)
-                and config.triton.codegen_upcast_to_fp32
-            ):
-                line += ".to(tl.float32)"
-                var_dtype = torch.float32
-
-            out = self.kernel.cse.generate(self.kernel.compute, line, dtype=var_dtype)
-            return out
-
-        return self.kernel.cse.generate(
-            self.kernel.compute, f"({self.fixed_inputs[name]})", dtype=torch.float32
-        )
-
-    def indirect_indexing(self, index_var: str, size, check, wrap_neg=True):
-        """Convert index variable to symbolic form."""
-        return sympy_index_symbol(str(index_var))
-
-    def store(
-        self, name: str, index: sympy.Expr, value: CSEVariable, mode: StoreMode = None
-    ) -> str:
-        """Currently only supports stores for atomic adds coming from scatter nodes
-        This is used by flex_attention's backwards grad for captured buffers, see
-        zeros_and_scatter lowering
-        """
-        assert self.mask is not None, (
-            "Mask is required for inner stores in modifications"
-        )
-        assert mode == "atomic_add", "Only atomic_add is supported for inner stores"
-
-        buf_name = self._add_kernel_input(name)
-        index_str = self._process_indexing(index)
-        index_str = f"tl.broadcast_to({index_str}, {value}.shape)"
-        store = f"tl.atomic_add({buf_name} + {index_str}, {value}, {self.mask}, sem='relaxed')"
-        return store
-
-    def _add_kernel_input(self, name: str):
-        """Add name as input to kernel and return input ref."""
-        return self.kernel.args.input(name)
-
-    def _process_indexing(self, index):
-        """Process and rename indexing, adding symbols as kernel inputs."""
-        return self.kernel.kexpr(self.kernel.rename_indexing(index))
-
-
-# Function name, followed by args and kwargs.
-RecordedEventsType = list[tuple[str, list[Any], dict[str, Any]]]
-
-
-class TritonTemplateKernel(TritonKernel):
-    def __init__(
-        self,
-        kernel_name,
-        input_nodes,
-        output_node,
-        defines,
-        num_stages,
-        num_warps,
-        grid_fn,
-        meta,
-        call_sizes,
-        num_consumer_groups=0,
-        num_buffers_warp_spec=0,
-        use_jit=False,
-        prefix_args=0,
-        suffix_args=0,
-        epilogue_fn=identity,
-        subgraphs: Optional[list[ir.ComputedBuffer]] = None,
-        workspace_arg: Optional[WorkspaceArg] = None,
-        prologue_loads_all_inputs=False,
-    ) -> None:
-        numel = sympy_product(output_node.get_size())
-        super().__init__(
-            {
-                "x": numel,
-                "r0_": sympy.S.One,
-            },
-            features=SIMDKernelFeatures([], numel),
-        )
-        self.input_nodes = input_nodes
-        self.output_node = output_node
-        self.named_input_nodes = {}  # type: ignore[var-annotated]
-        self.defines = defines
-        self.kernel_name = kernel_name
-        self.use_jit = use_jit
-        self.num_stages = num_stages
-        self.num_warps = num_warps
-        self.num_consumer_groups = num_consumer_groups
-        self.num_buffers_warp_spec = num_buffers_warp_spec
-        self.grid_fn = grid_fn
-        self.meta = meta
-        self.call_sizes = call_sizes
-        # for templates with fixed epilogues
-        self.prefix_args = prefix_args
-        self.suffix_args = suffix_args
-        self.epilogue_fn = epilogue_fn
-        self.render_hooks = {}  # type: ignore[var-annotated]
-        self.triton_meta: Optional[dict[str, object]] = None
-        # For Templated Attention this can be a list of ir.Subgraph
-        self.subgraphs: Optional[list[ir.ComputedBuffer]] = subgraphs
-
-        # Some templates use extra global memory as a workspace
-        self.workspace_arg = workspace_arg
-        if workspace_arg is not None:
-            self.args.workspace_args.append(workspace_arg)
-
-        # The following attributes (body, template_mask, output_val) are all
-        # used for triton kernel codegen.
-        # They are swapped onto the TritonTemplateKernel object by
-        # `set_subgraph_body`
-        self.subgraph_bodies: dict[str, SubgraphInfo] = {}
-
-        # input buffers which we are allowed to prologue fuse into
-        self.prologue_supported_inputs: OrderedSet[str] = OrderedSet()
-
-        # input buffers which we are fusing into
-        self.prologue_fused_inputs: OrderedSet[str] = OrderedSet()
-        # input buffers which we are fusing into, which preserve a zero mask
-        self.prologue_fused_inputs_preserve_zero: OrderedSet[str] = OrderedSet()
-
-        # The following attributes are all used for triton kernel codegen.
-        # They are swapped onto the TritonTemplateKernel object by
-        # `set_subgraph_body`
-        # NB: the names here must match the fields in SubgraphInfo
-        self.body: IndentedBuffer = FakeIndentedBuffer()
-        self.compute: IndentedBuffer = FakeIndentedBuffer()
-        self.indexing_code: IndentedBuffer = FakeIndentedBuffer()
-        self.loads: IndentedBuffer = FakeIndentedBuffer()
-        self.stores: IndentedBuffer = FakeIndentedBuffer()
-        self.template_mask: Optional[str] = None
-        self.template_out: Optional[str] = None
-        self.ops_handler: Optional[V.WrapperHandler] = None  # type: ignore[name-defined]
-
-        # Whe caching is enabled, the generated code is not dependent on the input nodes names, or
-        # symbolic sizes names.
-        # However, some of the variables returned by generate_and_load that are computed during the
-        # triton template expansions (code generation) are dependent on those.
-        # In order to cache the code generation and avoid redoing it for similar inputs that varies only by
-        # input names or symbol names, we do a record and replay method.
-        # During template expansions we record all function calls that change input_dependent_preserved_state
-        # and replay them on a cache hit to regenerate them.
-        self.cached_replay_events: Optional[RecordedEventsType] = None
-
-        # Update each time an input is marked frozen, used to replay the freezing of inputs on a cache hit.
-        self.frozen_layouts_cnt = 0
-
-        # When prologue_loads_all_inputs is true, prologue_supported_inputs is populated during def_kernel
-        # by adding all inputs.
-        self.prologue_loads_all_inputs = prologue_loads_all_inputs
-
-    def input_dependent_preserved_state(self) -> str:
-        # Not adding self.args.output_buffers on purpose. But we do not need to reproduce it on a cache hit.
-        # (never accessed).
-        return repr(
-            [
-                self.args.input_buffers,
-                self.args.sizevars,
-                self.args.workspace_args,
-                self.prologue_supported_inputs,
-                self.frozen_layouts_cnt,
-            ]
-        )
-
-    def record_input_dependent_tracked_event(self) -> Callable[..., Any]:
-        def decorator(fn) -> Callable[..., Any]:
-            def wrapper(*args, **kwargs) -> Any:
-                pre_state = self.input_dependent_preserved_state()
-                result = fn(*args, **kwargs)
-                post_state = self.input_dependent_preserved_state()
-                if pre_state != post_state:
-                    assert self.cached_replay_events is not None
-                    self.cached_replay_events.append((fn.__name__, [*args], {**kwargs}))
-                return result
-
-            return wrapper
-
-        return decorator
-
-    def replay_cached_events(self, events: RecordedEventsType) -> None:
-        for f, args, kwargs in events:
-            getattr(self, f)(*args, **kwargs)
-
-    @contextlib.contextmanager
-    def set_subgraph_body(self, body_name: str):
-        assert all(
-            hasattr(self, field.name) for field in dataclasses.fields(SubgraphInfo)
-        )
-        old_state = {
-            key.name: getattr(self, key.name)
-            for key in dataclasses.fields(SubgraphInfo)
-        }
-
-        assert body_name in self.subgraph_bodies, body_name
-
-        subgraph = self.subgraph_bodies[body_name]
-        for key, value in subgraph.to_dict().items():
-            if value is None and key in subgraph.only_copy_if_non_none_fields:
-                continue
-            setattr(self, key, value)
-
-        context = (
-            contextlib.nullcontext
-            if not self.ops_handler
-            else lambda: V.set_ops_handler(self.ops_handler(V.get_ops_handler()))
-        )
-        with context():  # type: ignore[operator]
-            yield
-        self.subgraph_bodies[body_name] = SubgraphInfo(
-            **{
-                key.name: getattr(self, key.name)
-                for key in dataclasses.fields(SubgraphInfo)
-            }
-        )
-        for key, value in old_state.items():
-            setattr(self, key, value)
-
-    @contextlib.contextmanager
-    def create_subgraph_body(self, body_name: str):
-        assert body_name not in self.subgraph_bodies
-        self.subgraph_bodies[body_name] = SubgraphInfo(
-            IndentedBuffer(),
-            None,
-            None,
-        )
-        with self.set_subgraph_body(body_name):
-            yield
-
-    def need_numel_args(self):
-        return False
-
-    def estimate_kernel_num_bytes(self):
-        """
-        Estimate the total number of bytes this kernel takes.
-        For in/out nodes, sizes are counted twice: once for reading and
-        once for writing.
-        """
-        ninplace_args = len(unique(self.args.inplace_buffers.values()))
-        num_bytes = []
-        for i, inp in enumerate(itertools.chain(self.input_nodes, (self.output_node,))):
-            size = V.graph.sizevars.size_hints(inp.get_size())
-            numel = functools.reduce(operator.mul, size, 1)
-            dtype_size = get_dtype_size(inp.get_dtype())
-            num_bytes.append(numel * dtype_size * (1 + int(i < ninplace_args)))
-        return sum(num_bytes)
-
-    def jit_lines(self):
-        if self.use_jit:
-            return "@triton.jit"
-
-        argdefs, _, signature, _ = self.args.python_argdefs()
-        triton_meta: dict[str, Any] = {
-            "signature": signature_to_meta(
-                signature,
-                size_dtype=self.index_dtype,
-                argdefs=argdefs,
-                is_template=True,
-            ),
-            "device": DeviceProperties.create(self.output_node.get_device()),
-            "constants": {},
-        }
-        triton_meta["configs"] = [config_of(signature)]
-        for arg_num in equal_1_arg_indices(signature):  # type: ignore[index]
-            triton_meta["constants"][signature[arg_num].name] = 1  # type: ignore[index,union-attr]
-        matrix_instr_nonkdim = self.meta.get("matrix_instr_nonkdim", None)
-        waves_per_eu = self.meta.get("waves_per_eu", None)
-        kpack = self.meta.get("kpack", None)
-        if matrix_instr_nonkdim:
-            triton_meta["matrix_instr_nonkdim"] = matrix_instr_nonkdim
-        if waves_per_eu:
-            triton_meta["waves_per_eu"] = waves_per_eu
-        if kpack:
-            triton_meta["kpack"] = kpack
-
-        self.triton_meta = triton_meta
-
-        inductor_meta = {
-            "kernel_name": str(Placeholder.DESCRIPTIVE_NAME),
-            **TritonKernel.inductor_meta_common(),
-            **FixedGrid.setup_grid_as_args(),
-        }
-        if config.profile_bandwidth or config.benchmark_kernel:
-            num_gb = self.estimate_kernel_num_bytes() / 1e9
-            inductor_meta["kernel_num_gb"] = num_gb
-
-        template_args = f"""
-            num_stages={self.num_stages},
-            num_warps={self.num_warps},
-            triton_meta={triton_meta!r},
-            inductor_meta={inductor_meta!r},
-        """
-
-        if HAS_WARP_SPEC:
-            template_args += f"""
-            num_consumer_groups={self.num_consumer_groups},
-            num_buffers_warp_spec={self.num_buffers_warp_spec},
-        """
-
-        return f"""
-            @triton_heuristics.template(
-                {template_args}
-            )
-            @triton.jit
-        """
-
-    def gen_argdefs(self):
-        def hook():
-            # python_argdefs() cannot be run until after the rest of the template lazily adds more args
-            arg_defs, *_ = self.args.python_argdefs()
-            return f"{', '.join(x.full_name() for x in arg_defs)}"
-
-        self.render_hooks["<ARGDEFS>"] = hook
-        return "<ARGDEFS>"
-
-    def gen_defines(self):
-        return self.defines
-
-    def def_kernel(self, *argnames):
-        """
-        Hook called from template code to generate function def and
-        needed args.
-        """
-        assert all(isinstance(x, str) for x in argnames)
-        renames = IndentedBuffer(initial_indent=1)
-
-        named_args = self.input_nodes[
-            self.prefix_args : len(self.input_nodes) - self.suffix_args
-        ]
-
-        assert len(argnames) == len(named_args), (
-            len(argnames),
-            len(named_args),
-            self.prefix_args,
-            len(self.input_nodes),
-        )
-
-        for input_node in self.input_nodes[: self.prefix_args]:
-            # get args in correct order
-            self.args.input(input_node.get_name())
-
-        for name, input_node in zip(argnames, named_args):
-            arg_name = f"arg_{name}"
-            self.named_input_nodes[name] = input_node
-            if input_node.get_name() in V.graph.removed_buffers:
-                continue
-            if input_node.get_name() in self.prologue_fused_inputs:
-                continue
-
-            self.args.input_buffers[input_node.get_name()] = arg_name
-
-        # The args may be duplicated, so renaming must be after args are de-duplicated.
-        for name in argnames:
-            input_node = self.named_input_nodes[name]
-            if self.prologue_loads_all_inputs:
-                self.prologue_supported_inputs.add(input_node.get_name())
-            if input_node.get_name() in V.graph.removed_buffers:
-                continue
-            if input_node.get_name() in self.prologue_fused_inputs:
-                continue
-
-            arg_name = self.args.input_buffers[input_node.get_name()]
-            if input_node.get_layout().offset == 0:
-                renames.writeline(f"{name} = {arg_name}")
-            else:
-                offset = texpr(self.rename_indexing(input_node.get_layout().offset))
-                renames.writeline(f"{name} = {arg_name} + {offset}")
-
-        for input_node in self.input_nodes[len(self.input_nodes) - self.suffix_args :]:
-            # get args in correct order
-            if input_node.get_name() in V.graph.removed_buffers:
-                continue
-            if input_node.get_name() in self.prologue_fused_inputs:
-                continue
-
-            self.args.input(input_node.get_name())
-
-        def hook():
-            # python_argdefs() cannot be run until after the rest of the template lazily adds more args
-            arg_defs, *_ = self.args.python_argdefs()
-            code = IndentedBuffer()
-            code.splice(gen_common_triton_imports())
-            code.splice(self.jit_lines())
-            code.writeline(
-                f"def {self.kernel_name}({', '.join(x.full_name() for x in arg_defs)}):"
-            )
-            with code.indent():
-                code.splice(self.defines)
-                code.splice(renames.getvalue())
-            return code.getvalue()
-
-        assert "<DEF_KERNEL>" not in self.render_hooks
-        self.render_hooks["<DEF_KERNEL>"] = hook
-        return "<DEF_KERNEL>"
-
-    def size(self, name: str, index: int):
-        """
-        Hook called from template code to get the size of an arg.
-        Will add needed args to pass it in if it is dynamic.
-        """
-        assert isinstance(index, int)
-        if name is None:
-            val = self.output_node.get_size()[index]
-        else:
-            assert isinstance(name, str)
-            val = self.named_input_nodes[name].get_size()[index]
-        return texpr(self.rename_indexing(val))
-
-    def stride(self, name, index=None):
-        """
-        Hook called from template code to get the stride of an arg.
-        Will add needed args to pass it in if it is dynamic.
-        """
-        if name is None:
-            val = self.output_node.get_stride()
-        else:
-            assert isinstance(name, str)
-            val = self.named_input_nodes[name].get_stride()
-
-        if isinstance(index, int):
-            return texpr(self.rename_indexing(val[index]))
-        return ", ".join([texpr(self.rename_indexing(i)) for i in val])
-
-    def _get_subgraph(self, subgraph_number: int):
-        assert isinstance(subgraph_number, int)
-        assert isinstance(self.subgraphs, list)
-        assert subgraph_number < len(self.subgraphs), (
-            f"Invalid subgraph number provided to create_modification, {subgraph_number} must be < {len(self.subgraphs)}"
-        )
-        assert self.body.getvalue() == "", (
-            "Body should be clear before adding a modification"
-        )
-        return self.subgraphs[subgraph_number]
-
-    def _handle_scatter_graph(self, scatter_graph):
-        """Handle processing for a single scatter graph.
-
-        Args:
-            scatter_graph: The scatter graph to process
-        """
-        assert isinstance(scatter_graph, ir.ComputedBuffer), (
-            f"scatter_graph must be an instance of ComputeBuffer but got {type(scatter_graph)}"
-        )
-
-        def contiguous_strides(x):
-            # We always create a fresh contiguous grad for scattering into
-            return sum(
-                x_i * stride for x_i, stride in zip(x, scatter_graph.get_stride())
-            )
-
-        return scatter_graph.data.store_output(  # type: ignore[attr-defined]
-            scatter_graph.name, contiguous_strides, []
-        )
-
-    def modification(
-        self,
-        subgraph_number: int,
-        output_name: Optional[str],
-        mask: Optional[str] = None,
-        **fixed_inputs,
-    ) -> str:
-        """This creates a modification function for a subgraph.
-        To use this inside a template, the first argument should specify which subgraph to codegen for
-
-        Args:
-            subgraph_number (int): The index of the subgraph in self.subgraphs
-            output_name (Optional[str]): The name of the output variable to store the result in
-            mask (Optional[str]): An optional mask to use for the store operation. If provided, this mask
-                will be applied to the store.
-        """
-        num = 0
-        out = None
-        scatters = []
-        while f"mod_{subgraph_number}_{num}" in self.subgraph_bodies:
-            num += 1
-        with self.create_subgraph_body(f"mod_{subgraph_number}_{num}"):
-            subgraph = self._get_subgraph(subgraph_number)
-            modification_handler = ModificationWrapper(
-                self, subgraph_number, fixed_inputs, mask
-            )
-            with V.set_ops_handler(modification_handler):
-                assert isinstance(subgraph, (ir.ComputedBuffer, list)), (
-                    f"Expected the subgraph to be a ComputedBuffer or a List[ComputedBuffer], got {type(subgraph)}"
-                )
-                # Handle scatter stores
-                if isinstance(subgraph, list):
-                    for scatter_graph in subgraph:
-                        scatters.append(self._handle_scatter_graph(scatter_graph))
-                elif isinstance(subgraph.data, ir.InputBuffer):
-                    out = subgraph.data.make_loader()(())
-                else:
-                    out = subgraph.data.inner_fn(())
-
-            self.codegen_body()
-            if output_name is not None:
-                assert isinstance(output_name, str)
-                assert out is not None
-                self.body.writeline(f"{output_name} = {out.value}")
-            else:
-                assert out is None
-                for scatter in scatters:
-                    self.body.writeline(str(scatter))
-
-            body_val = self.body.getvalue()
-            self.cse.invalidate(OrderedSet())
-            return body_val
-
-    def load_input(
-        self,
-        input_name: str,
-        output_name: str,
-        indices: Union[list[Any], tuple[Any]],
-        mask: Optional[str] = None,
-        other: Optional[Union[float, int]] = 0.0,
-        indent_width: int = 4,
-    ):
-        """Loads an input and applies any necessary preprocessing or masking.
-
-        Args:
-            input_name (str): The name of the input to load.
-            indices (Union[List, Tuple]): The index for each dimension of the input.
-            val (str): The name of the variable to store the loaded value.
-            mask (Optional[str]): An optional mask to use for the load operation.
-            other (Optional[Union[float, int]]): The value to use for masked elements. Default is 0.0.
-            indent_width (int): The number of spaces to use for indentation.
-        """
-
-        input_node = self.named_input_nodes[input_name]
-        if not self.prologue_loads_all_inputs:
-            self.prologue_supported_inputs.add(input_node.get_name())
-
-        tilings = (sympy_product(input_node.get_size()), sympy.Integer(1))
-        groups = {
-            "x": tilings[0],
-            "r0_": tilings[1],
-        }
-
-        range_trees = self.construct_range_trees(
-            pid_cache=None,
-            inside_reduction=False,
-            is_reduction=False,
-            numels=groups,
-            no_x_dim=False,
-        )
-        load_code = None
-
-        with self.create_subgraph_body(f"<LOAD_INPUT_{input_name}>"):
-            assert isinstance(indices, (list, tuple))
-            assert isinstance(output_name, str)
-            assert isinstance(mask, (str, type(None)))
-            self.range_trees = range_trees
-            self.numels = {k: V.graph.sizevars.simplify(v) for k, v in groups.items()}
-            indices = list(map(OpOverrides.paren, indices))
-            index_symbols = [sympy.Symbol(x, integer=True) for x in indices]
-
-            lengths = [V.graph.sizevars.simplify(s) for s in input_node.get_size()]
-            assert len(indices) == len(lengths)
-
-            index_symbols = [sympy.Symbol(x, integer=True) for x in indices]
-            assert len(indices) == len(lengths)
-
-            # glue to make generated code use same indexing from template
-
-            # TODO (from reviewers as well)
-            # in codegen_template,
-            # prologue_node.codegen(kernel.split_and_set_ranges(prologue_node.get_ranges()))
-            # the ranges need to reflect the group of the prologue input or it will error
-            # not sure if there is any difference between original range_tree_entry in
-            # and new one from correct lengths/groups... both actually seem to work
-            for name, range_tree_entry in zip(
-                indices, self.range_trees[0].construct_entries(lengths)
-            ):
-                range_tree_entry.set_name(name)
-            contiguous_index = sympy_dot(
-                ir.FlexibleLayout.contiguous_strides(lengths), index_symbols
-            )
-            contiguous_index = self.rename_indexing(contiguous_index)
-            self.body.writeline("xindex = " + texpr(contiguous_index))
-
-            xindex_range_root = self.range_trees[0].lookup(
-                sympy.Integer(1), sympy_product(lengths)
-            )
-            xindex_range_root.set_name("xindex")
-
-            # Note - ["None" override_mask]
-            # MM Templates work by taking out of bounds index values and wrapping them around to 0
-            # so that no mask is required on the load: offs_a_m = `rm % M`
-            # We should to override the mask to be "None" instead of inheriting the mask that would
-            # have been loaded otherwise.
-            # We are using "None" for clarity in output code, but
-            # we could alternatively emit `xmask = tl.full([xindex.shape], True, tl.int1)`
-            self.template_mask = mask if mask is not None else "None"
-            self.template_out = "xindex"
-            self.template_indices = indices
-            self.named_input_nodes[input_name].data.freeze_layout()
-            self.cse.invalidate(OrderedSet())
-
-            template_mask = self.template_mask
-
-            class StoreOutputSubstitution(V.WrapperHandler):  # type: ignore[name-defined]
-                name = "StoreOutputSubstitution"
-
-                def store(
-                    self,
-                    name: str,
-                    index: sympy.Expr,
-                    value: "CSEVariable",
-                    mode: "StoreMode" = None,
-                ):
-                    V.kernel.store_buffer_names.add(name)
-                    V.kernel.cse.store_cache[name] = value
-                    if name in V.kernel.prologue_fused_inputs:
-                        # We load masked out values with 0, then apply a prologue.
-                        # The masked out values may not necessariliy be 0 any more
-                        # so we need to reapply the mask.
-                        value_dtype = value.dtype
-                        value_str = str(value)
-                        if template_mask != "None" and (
-                            name not in V.kernel.prologue_fused_inputs_preserve_zero
-                            or other != 0
-                        ):
-                            value_str = (
-                                f"tl.where({template_mask}, {value_str}, {other})"
-                            )
-
-                        if value_dtype != V.graph.get_buffer(name).dtype:
-                            value_str = f"{value_str}.to({triton_type(V.graph.get_buffer(name).dtype)})"
-
-                        # TODO: we should have intermediary var shapes
-                        V.kernel.compute.writeline(
-                            f"{output_name} = {value_str}.broadcast_to(xindex.shape)"
-                        )
-
-            self.ops_handler = StoreOutputSubstitution
-
-            input_node = self.named_input_nodes[input_name]
-            output_index = input_node.make_indexer()(index_symbols)
-
-            # in def_kernel above we define the inputs with the storage offset adjusted
-            # creating the load in input_node.make_indexer() will also adjust by storage offset
-            # so subtract here to not double increment
-            if not V.graph.sizevars.statically_known_equals(
-                input_node.layout.offset, 0
-            ):
-                output_index = output_index - self.rename_indexing(
-                    input_node.get_layout().offset
-                )
-
-            output_index = self.rename_indexing(output_index)
-
-            if output_index == contiguous_index:
-                output_index_str = "xindex"
-            else:
-                out_indexing = self.indexing(
-                    output_index,
-                    copy_shape=self.template_out,
-                    override_mask=self.template_mask,
-                )
-                from .codegen.triton import IndexingOptions
-
-                assert isinstance(out_indexing, IndexingOptions)
-                output_index_str = (
-                    f"({out_indexing.index_str}).broadcast_to(xindex.shape)"
-                )
-
-            # Generate load code
-            load_code = f"{output_name} = tl.load({input_name} + ({output_index_str})"
-
-            if mask:
-                load_code += f", mask={mask}, other={other})"
-            else:
-                load_code += ")"
-
-        hook_key = f"<LOAD_INPUT_{input_name}>"
-
-        def hook():
-            with self.set_subgraph_body(hook_key):
-                self.cse.invalidate(OrderedSet())
-                self.codegen_body()
-                self.cse.invalidate(OrderedSet())
-                if input_node.get_name() not in self.prologue_fused_inputs:
-                    assert load_code is not None
-                    self.body.writeline(load_code)
-
-                return textwrap.indent(self.body.getvalue(), " " * indent_width).strip()
-
-        assert hook_key not in self.render_hooks
-        self.render_hooks[hook_key] = hook
-        return hook_key
-
-    def store_output(
-        self,
-        indices: Union[list[Any], tuple[Any]],
-        val: str,
-        mask: Optional[str] = None,
-        indent_width: int = 4,
-    ):
-        """Stores the final output and appends any epilogue fusions if the buffer hasn't been optimized away.
-
-        Args:
-            indices (Union[List, Tuple]): The index for each dimension of the output. The dot product of
-                these indices and output strides must match `val`.
-            val (str): The value to store.
-            mask (Optional[str]): An optional mask to use for the store operation. If provided, this mask
-                will be applied to the store.
-            indent_width (int): The number of spaces to use for indentation. This is used when the call to
-                store_output is indented in the kernel definition.
-        """
-        with self.create_subgraph_body("<STORE_OUTPUT>"):
-            assert isinstance(indices, (list, tuple))
-            assert isinstance(val, str)
-            assert isinstance(mask, (str, type(None)))
-            assert self.template_mask is None
-            indices = list(map(OpOverrides.paren, indices))
-            index_symbols = [sympy.Symbol(x, integer=True) for x in indices]
-            lengths = [
-                V.graph.sizevars.simplify(s) for s in self.output_node.get_size()
-            ]
-            assert len(indices) == len(lengths)
-
-            # glue to make generated code use same indexing from template
-            for name, range_tree_entry in zip(
-                indices, self.range_trees[0].construct_entries(lengths)
-            ):
-                range_tree_entry.set_name(name)
-            contiguous_index = sympy_dot(
-                ir.FlexibleLayout.contiguous_strides(lengths), index_symbols
-            )
-            contiguous_index = self.rename_indexing(contiguous_index)
-            self.body.writeline("xindex = " + texpr(contiguous_index))
-            self.range_trees[0].lookup(sympy.S.One, sympy_product(lengths)).set_name(
-                "xindex"
-            )
-            self.template_mask = mask
-            self.template_out = val
-            self.template_indices = indices
-            output_index = self.output_node.get_layout().make_indexer()(index_symbols)
-            output_index = self.rename_indexing(output_index)
-            if output_index == contiguous_index:
-                output_index = sympy.Symbol("xindex", integer=True)
-
-            acc_dtype = (
-                triton_type_to_torch(self.meta["ACC_TYPE"])
-                if "ACC_TYPE" in self.meta
-                else torch.float32
-            )
-            epilogue_args = [V.kernel.cse.namedvar(val, dtype=acc_dtype)]
-            for input_node in itertools.chain(
-                self.input_nodes[: self.prefix_args],
-                self.input_nodes[len(self.input_nodes) - self.suffix_args :],
-            ):
-                input_node.freeze_layout()
-                epilogue_args.append(input_node.make_loader()(index_symbols))
-                # We update frozen_layouts_cnt in order to replay this function on a cache hit.
-                self.frozen_layouts_cnt += 1
-
-            V.ops.store(
-                self.output_node.get_name(),
-                output_index,
-                self.epilogue_fn(*epilogue_args),
-            )
-            self.codegen_body()
-
-        def hook():
-            # more stuff might have been added since the codegen_body above
-            self.codegen_body()
-            self.cse.invalidate(OrderedSet())
-
-            return textwrap.indent(self.body.getvalue(), " " * indent_width).strip()
-
-        assert "<STORE_OUTPUT>" not in self.render_hooks
-        self.render_hooks["<STORE_OUTPUT>"] = hook
-        return "<STORE_OUTPUT>"
-
-    def render(self, template, kwargs, record_input_dependent_tracked_event=False):
-        if record_input_dependent_tracked_event:
-            self.cached_replay_events = []
-
-        template_env = {
-            fn.__name__: self.record_input_dependent_tracked_event()(fn)
-            if record_input_dependent_tracked_event
-            else fn
-            for fn in [
-                self.def_kernel,
-                self.size,
-                self.stride,
-                self.store_output,
-                self.load_input,
-                self.make_load,
-                self.modification,
-                self.gen_argdefs,
-                self.gen_defines,
-            ]
-        }
-        return PartialRender(
-            template.render(**template_env, **kwargs),
-            self.render_hooks,
-        )
-
-    def make_load(self, name, indices, mask):
-        """
-        Optional helper called from template code to generate the code
-        needed to load from an tensor.
-        """
-        assert isinstance(indices, (list, tuple))
-        assert isinstance(name, str)
-        assert isinstance(mask, str)
-        stride = self.named_input_nodes[name].get_stride()
-        indices = list(map(OpOverrides.paren, indices))
-        assert len(indices) == len(stride)
-        index = " + ".join(
-            f"{texpr(self.rename_indexing(s))} * {i}" for s, i in zip(stride, indices)
-        )
-        return f"tl.load({name} + ({index}), {mask}, other=0.0)"
-
-    def indexing(
-        self,
-        index: sympy.Expr,
-        *,
-        dense_indexing=False,
-        copy_shape=None,
-        override_mask=None,
-        block_ptr=False,
-    ):
-        """
-        Override the default indexing to use our custom mask and force
-        dense indexing.
-        """
-        return super().indexing(
-            index,
-            dense_indexing=False,
-            # We pass template_out as the shape to broadcast the indexing to as
-            # the mask might be broadcast to the output shape
-            copy_shape=self.template_out,
-            override_mask=self.template_mask,
-            block_ptr=block_ptr,
-        )
-
-    def codegen_range_tree(self):
-        pass  # ignore default codegen
-
-    def call_kernel(self, name: str, node: Optional[ir.IRNode] = None):
-        wrapper = V.graph.wrapper_code
-        _, call_args, _, arg_types = self.args.python_argdefs()
-
-        grid_args = ()
-        if isinstance(self.grid_fn, SymbolicGridFn):
-            grid_args = self.grid_fn.sympy_call(*self.call_sizes, self.meta)
-        elif all(isinstance(x, (int, sympy.Integer)) for x in self.call_sizes):
-            grid_args = self.grid_fn(*map(int, self.call_sizes), self.meta)
-        else:
-            assert not V.graph.cpp_wrapper, "cpp_wrapper requires SymbolicGridFn"
-            wrapper.add_import_once(f"import {self.grid_fn.__module__}")
-            meta = wrapper.add_meta_once(self.meta)
-            fn_name = f"{self.grid_fn.__module__}.{self.grid_fn.__name__}"
-            call_args.append(
-                f"*{fn_name}({', '.join(map(pexpr, self.call_sizes))}, {meta})"
-            )
-            arg_types.append(None)
-        assert len(grid_args) in (0, 3), "grid_fn should return 3 values"
-        call_args.extend(grid_args)
-        arg_types.extend(map(type, grid_args))
-
-        if self.workspace_arg is not None:
-            wrapper.generate_workspace_allocation(self.workspace_arg)
-        wrapper.generate_kernel_call(
-            name,
-            call_args,
-            arg_types=arg_types,
-            triton_meta=self.triton_meta,
-            triton=True,
-        )
-        if self.workspace_arg is not None:
-            wrapper.generate_workspace_deallocation(self.workspace_arg)
-
-    def kernel_benchmark_extra_args(self) -> list[str]:
-        return [
-            str(x)
-            for x in self.grid_fn(
-                *V.graph.sizevars.size_hints(self.call_sizes), self.meta
-            )
-        ]
-
-
 @functools.cache
 def _jinja2_env():
     try:
@@ -1134,566 +152,6 @@ def _jinja2_env():
         )
     except ImportError:
         return None
-
-
-class GenerateAndLoadResult(NamedTuple):
-    """
-    Return type of TritonTemplate.generate_and_load.
-    """
-
-    mod: ModuleType
-    extra: str
-    input_call_args: tuple[str, ...]
-    prologue_supported_inputs: OrderedSet[str]
-    kernel_args_sizevars_keys: tuple[sympy.Expr]
-    kernel_options: dict[str, Any]
-
-
-class GeneratedCodeCacheEntry(NamedTuple):
-    code: str
-    extra: str
-    events: list[Any]
-
-
-class GeneratedCodeCache:
-    """
-    Cache for generated code. The cache key is a string representation of the input nodes,
-    number of stages, number of warps, and call sizes. The cache value is a tuple of the
-    generated code, extra code, and events.
-    """
-
-    def __init__(self, *args, **kwargs):
-        self._cache: dict[str, GeneratedCodeCacheEntry] = {}
-
-    def cache_clear(self) -> None:
-        self._cache.clear()
-
-    def __repr__(self):
-        return repr(self._cache)
-
-    def make_key(
-        self,
-        input_nodes: tuple[ir.IRNode],
-        num_stages: int,
-        num_warps: int,
-        call_sizes: list[sympy.core.symbol.Symbol],
-        prefix_args: int,
-        suffix_args: int,
-        epilogue_fn: Optional[Callable[..., Any]],
-        epilogue_fn_hash: Optional[str],
-        subgraphs: Optional[list[ir.Buffer]],  # has to be none to cache
-        workspace_arg: Optional[WorkspaceArg],  # has to be none to cache
-        layout: ir.Layout,
-        num_consumer_groups: int,
-        num_buffers_warp_spec: int,
-        kwargs: dict[str, Any],
-    ) -> Optional[str]:
-        def layout_key(layout: ir.Layout) -> str:
-            assert not isinstance(layout, ir.FlexibleLayout)
-            return repr(
-                [
-                    layout.size,
-                    layout.stride,
-                    layout.dtype,
-                    layout.device,
-                    layout.offset,
-                ]
-            )
-
-        def has_flexible_layout() -> bool:
-            if isinstance(layout, ir.FlexibleLayout):
-                return True
-
-            for input in input_nodes:
-                if isinstance(input.get_layout(), ir.FlexibleLayout):
-                    return True
-            return False
-
-        if epilogue_fn is identity:
-            assert epilogue_fn_hash is None
-            epilogue_fn_hash = "identity"
-
-        # we do not cache under those conditions right now.
-        if (
-            has_flexible_layout()
-            or subgraphs is not None
-            or workspace_arg is not None
-            or epilogue_fn_hash is None
-        ):
-            return None
-
-        return repr(
-            {
-                "input_nodes": [
-                    layout_key(input.get_layout()) for input in input_nodes
-                ],
-                "num_stages": num_stages,
-                "num_warps": num_warps,
-                "prefix_args": prefix_args,
-                "suffix_args": suffix_args,
-                "call_sizes": call_sizes,
-                "layout": layout_key(layout),
-                "num_consumer_groups": num_consumer_groups,
-                "num_buffers_warp_spec": num_buffers_warp_spec,
-                "epilogue_fn_hash": epilogue_fn_hash,
-                "kwargs": kwargs,
-            }
-        )
-
-    def get_entry(self, cache_key: Optional[str]) -> Optional[GeneratedCodeCacheEntry]:
-        if cache_key is None:
-            return None
-
-        entry = self._cache.get(cache_key, None)
-        if entry is None:
-            torch._dynamo.utils.counters["inductor"]["generated_module_cache_miss"] += 1
-        else:
-            torch._dynamo.utils.counters["inductor"]["generated_module_cache_hit"] += 1
-        return entry
-
-    def put_entry(
-        self,
-        cache_key: Optional[str],
-        code: str,
-        extra: str,
-        events: list[Any],
-    ) -> None:
-        if cache_key is None:
-            return
-        entry = GeneratedCodeCacheEntry(code, extra, events)
-        self._cache.update({cache_key: entry})
-
-
-class TritonTemplate(KernelTemplate):
-    """
-    A Triton template is a template that can be used to generate a Triton kernel.
-    """
-
-    # Allow subclasses to override the kernel type
-    kernel_type: type[Any] = TritonTemplateKernel
-    index_counter = itertools.count()
-    all_templates: dict[str, "TritonTemplate"] = {}
-
-    def __init__(
-        self,
-        name: str,
-        grid: Any,
-        source: str,
-        debug=False,
-        cache_codegen_enabled_for_template=False,
-        prologue_loads_all_inputs=False,
-    ) -> None:
-        super().__init__(name)
-        self.grid = grid
-        self.template = self._template_from_string(source)
-        assert name not in self.all_templates, "duplicate template name"
-        TritonTemplate.all_templates[name] = self
-        self.debug = debug
-        self._cache_codegen_enabled_for_template = cache_codegen_enabled_for_template
-        self._generated_code_cache: GeneratedCodeCache = GeneratedCodeCache()
-        clear_on_fresh_cache(self._generated_code_cache)
-        # When prologue_loads_all_inputs is true, prologue_supported_inputs is populated during def_kernel
-        # by adding all inputs.
-        self.prologue_loads_all_inputs = prologue_loads_all_inputs
-
-    # When this flag is on, we ensure that the cached results and the generated result if cache
-    # was not used are the same.
-    test_cache = False
-
-    def maybe_append_choice(
-        self, choices: list[Any], **kwargs: Any
-    ) -> Optional[NotImplementedError]:
-        """
-        Maybe generates a new ChoiceCaller and appends it into existing choices.
-        Returns None if success, otherwise returns the error.
-
-        choices: A list of ChoiceCallers.
-        kwargs: Additional kwargs to be passed to self.generate() to generate a new ChoiceCaller.
-        """
-
-        try:
-            choices.append(self.generate(generate_with_caching=True, **kwargs))
-            return None
-        except NotImplementedError as e:
-            log.info(
-                "Cannot Append Choice: %s. KernelTemplate type is %s",
-                e,
-                type(self),
-                stack_info=log.getEffectiveLevel() < logging.INFO,
-            )
-            return e
-
-    # NOTE: MAKE SURE THAT ANY ARGUMENT ADDED TO THIS FUNCTION IS PROPERLY HANDLED IN _generated_code_cache.make_key.
-    def generate_and_load(
-        self,
-        input_nodes: tuple[ir.IRNode],
-        num_stages: int,
-        num_warps: int,
-        call_sizes: list[sympy.core.symbol.Symbol],
-        prefix_args: int,
-        suffix_args: int,
-        epilogue_fn: Optional[Callable[..., Any]],
-        epilogue_fn_hash: Optional[str],
-        subgraphs: Optional[list[ir.Buffer]],
-        workspace_arg: Optional[WorkspaceArg],
-        num_consumer_groups: int,
-        num_buffers_warp_spec: int,
-        layout: ir.Layout,
-        kwargs: dict[str, Any],
-        generate_with_caching,
-    ) -> Optional[GenerateAndLoadResult]:
-        """Generate the python code and load it into the current process"""
-        caching_enabled = (
-            generate_with_caching
-            and torch._inductor.config.enable_caching_generated_triton_templates
-        )
-
-        cache_key = None
-        if caching_enabled:
-            cache_key = self._generated_code_cache.make_key(
-                input_nodes,
-                num_stages,
-                num_warps,
-                call_sizes,
-                prefix_args,
-                suffix_args,
-                epilogue_fn,
-                epilogue_fn_hash,
-                subgraphs,
-                workspace_arg,
-                layout,
-                num_consumer_groups,
-                num_buffers_warp_spec,
-                kwargs,
-            )
-
-        assert self.template, "requires jinja2"
-        defines = StringIO()
-
-        for name, val in kwargs.items():
-            defines.write(f"{name} : tl.constexpr = {val}\n")
-        defines = defines.getvalue()
-
-        fake_out = ir.Buffer(name="buf_out", layout=layout)
-        kernel_name = f"triton_{self.name}"
-
-        numel = sympy_product(layout.size)
-        buffers = itertools.chain(input_nodes, (fake_out,))
-        if not TritonScheduling.can_use_32bit_indexing(numel, buffers):
-            raise NotImplementedError(
-                "64-bit indexing is not yet implemented for triton templates"
-            )
-
-        kernel_options = {
-            "input_nodes": input_nodes,
-            "defines": defines,
-            "num_stages": num_stages,
-            "num_warps": num_warps,
-            "grid_fn": self.grid,
-            "meta": kwargs,
-            "call_sizes": call_sizes,
-            "prefix_args": prefix_args,
-            "suffix_args": suffix_args,
-            "epilogue_fn": epilogue_fn,
-            "subgraphs": subgraphs,
-            "prologue_loads_all_inputs": self.prologue_loads_all_inputs,
-        }
-
-        if HAS_WARP_SPEC:
-            kernel_options.update(
-                {
-                    "num_consumer_groups": num_consumer_groups,
-                    "num_buffers_warp_spec": num_buffers_warp_spec,
-                }
-            )
-
-        def make_kernel():
-            return self.kernel_type(
-                kernel_name=kernel_name,
-                output_node=fake_out,
-                workspace_arg=workspace_arg,
-                use_jit=False,
-                **kernel_options,
-            )
-
-        def generate_code(kernel) -> Optional[tuple[str, str]]:
-            def make_extra() -> str:
-                extra_parts = [
-                    f"{kwarg}={repr(kwargs[kwarg])}" for kwarg in sorted(kwargs.keys())
-                ]
-
-                extra_parts.extend(
-                    [
-                        f"num_stages={num_stages}",
-                        f"num_warps={num_warps}",
-                    ]
-                )
-                if HAS_WARP_SPEC:
-                    extra_parts.extend(
-                        [
-                            f"num_consumer_groups={num_consumer_groups}",
-                            f"num_buffers_warp_spec={num_buffers_warp_spec}",
-                        ]
-                    )
-                extra = "-".join(extra_parts) + "-"
-                return extra
-
-            try:
-                template = kernel.render(self.template, kwargs, caching_enabled)
-                with kernel.set_subgraph_body("<STORE_OUTPUT>"):
-                    code = template.finalize_all()
-            except ZeroDivisionError:
-                # TODO(nmacchioni): fix sympy division by zero
-                return None
-            if self.debug:
-                print("Generated Code:\n", code)
-
-            extra = make_extra()
-            return code, extra
-
-        def maybe_test_cache(code: str, extra: str, kernel):
-            if self.test_cache or self.debug:
-                with (
-                    patch.object(V.graph, "get_dtype", self._fake_get_dtype(fake_out)),
-                    V.graph.set_current_device(layout.device),
-                    make_kernel() as kernel_test,
-                ):
-                    result2 = generate_code(kernel_test)
-                    assert result2 is not None
-                    code_test, extra_test = result2
-                    assert (
-                        code == code_test
-                        and extra == extra_test
-                        and kernel.args.input_buffers == kernel_test.args.input_buffers
-                        and kernel.prologue_supported_inputs
-                        == kernel_test.prologue_supported_inputs
-                        and kernel.args.sizevars == kernel_test.args.sizevars
-                    ), "Generated code cache results in wrong output"
-
-        # Generate code, extra.
-        code: Optional[str] = None
-        extra: Optional[str] = None
-        with (
-            patch.object(V.graph, "get_dtype", self._fake_get_dtype(fake_out)),
-            V.graph.set_current_device(layout.device),
-            make_kernel() as kernel,
-        ):
-            cache_entry = self._generated_code_cache.get_entry(cache_key)
-            cache_hit = False
-
-            if cache_entry is not None:
-                code, extra, events = cache_entry
-                kernel.replay_cached_events(events)
-                cache_hit = True
-
-            else:
-                result = generate_code(kernel)
-                if result is None:  # happens at ZeroDivisionError:
-                    return None
-                code, extra = result
-                self._generated_code_cache.put_entry(
-                    cache_key, code, extra, kernel.cached_replay_events
-                )
-
-        assert code is not None and extra is not None
-
-        mod = PyCodeCache.load(code, extra)
-
-        input_call_args = tuple(kernel.args.input_buffers.keys())
-        prologue_supported_inputs = kernel.prologue_supported_inputs.copy()
-        kernel_args_sizevars_keys = tuple(kernel.args.sizevars.keys())
-
-        if cache_hit:
-            maybe_test_cache(code, extra, kernel)
-
-        return GenerateAndLoadResult(
-            mod,
-            extra,
-            input_call_args,
-            prologue_supported_inputs,
-            kernel_args_sizevars_keys,
-            kernel_options,
-        )
-
-    def generate(  # type: ignore[override]
-        self,
-        input_nodes: tuple[ir.IRNode],
-        layout: ir.Layout,
-        num_stages: int,
-        num_warps: int,
-        num_consumer_groups: int = 0,
-        num_buffers_warp_spec: int = 0,
-        prefix_args: int = 0,
-        suffix_args: int = 0,
-        epilogue_fn: Optional[Callable[..., Any]] = identity,
-        epilogue_fn_hash: Optional[str] = None,
-        subgraphs: Optional[list[ir.Buffer]] = None,
-        mutated_inputs: Optional[list[ir.IRNode]] = None,
-        call_sizes: Optional[list[sympy.core.symbol.Symbol]] = None,
-        workspace_arg: Optional[WorkspaceArg] = None,
-        generate_with_caching=False,
-        **kwargs,
-    ):
-        """This function generates a TritonTemplateCaller
-
-        Args:
-            input_nodes: List of input nodes
-            layout: Output layout
-            num_stages: Number of stages for triton launch
-            num_warps: Number of warps for triton launch
-            prefix_args: Number of input nodes to be passed as arguments
-            suffix_args: Number of input nodes to be passed as arguments
-            epilogue_fn: Optional epilogue function to be called on the output
-            subgraphs: Optional subgraphs to be passed as arguments, these will be inlined
-                into the triton template string
-            mutated_inputs: Optional list of input nodes that are mutated by the kernel, this is helpful
-                if you need to return multiple outputs. You can pass them as inputs and mark them as
-                being mutated by the kernel.
-        """
-        # HACK: Triton currently breaks if TF32 floats are requested, but the CUDA
-        # capability doesn't support them.  This is a bug in Triton, but for now we'll
-        # patch around it here.  See https://github.com/triton-lang/triton/issues/3011
-        # for one example issue with this problem.
-        if torch.cuda.is_available() and not torch.cuda.is_tf32_supported():
-            kwargs["ALLOW_TF32"] = "False"
-
-        if call_sizes is None:
-            call_sizes = layout.size
-
-        result = self.generate_and_load(
-            input_nodes,
-            num_stages,
-            num_warps,
-            call_sizes,
-            prefix_args,
-            suffix_args,
-            epilogue_fn,
-            epilogue_fn_hash,
-            subgraphs,
-            workspace_arg,
-            num_consumer_groups,
-            num_buffers_warp_spec,
-            layout,
-            kwargs,
-            generate_with_caching and self._cache_codegen_enabled_for_template,
-        )
-
-        # May happen as result of dev by 0.
-        if result is None:
-            return None
-
-        # We expect the input_buffer order to be [*input_nodes, *captured_buffers]
-        expected_input_args = tuple(unique(x.get_name() for x in input_nodes))
-        assert (
-            result.input_call_args[: len(expected_input_args)] == expected_input_args
-        ), (
-            result.input_call_args,
-            expected_input_args,
-        )
-
-        full_input_nodes = tuple(
-            [V.graph.get_buffer(k) for k in result.input_call_args]
-        )
-        extra_args = V.graph.sizevars.size_hints(
-            map(sympy.expand, result.kernel_args_sizevars_keys),
-            fallback=config.unbacked_symint_fallback,
-        )
-
-        kernel_hash_name = f"triton_{self.name}_{next(self.index_counter)}"
-
-        workspace_args = []
-        if workspace_arg is not None:
-            # Create workspace tensor
-            workspace_size = workspace_arg.count
-            workspace_tensor = torch.empty_strided(
-                (workspace_size,),
-                (1,),
-                dtype=torch.uint8,
-                device=layout.device.type,
-            )
-
-            # Handle zero initialization if needed
-            if workspace_arg.zero_mode != WorkspaceZeroMode.UNINITIALIZED:
-                workspace_tensor.zero_()
-
-            workspace_args.append(workspace_tensor)
-
-        options = result.kernel_options
-
-        def make_kernel_render(out_node):
-            assert result is not None
-            kernel = self.kernel_type(
-                kernel_name=str(Placeholder.KERNEL_NAME),
-                output_node=out_node,
-                workspace_arg=workspace_arg,
-                use_jit=False,
-                **options,
-            )
-            render = functools.partial(
-                kernel.render,
-                self.template,
-                kwargs,
-            )
-            return kernel, render
-
-        # create the BenchmarkRequest
-        assert result.mod.__file__ is not None
-        grid = self.grid(
-            *V.graph.sizevars.size_hints(
-                call_sizes,
-                fallback=config.unbacked_symint_fallback,
-            ),
-            kwargs,
-        )
-        bmreq_cls: type[TritonBenchmarkRequest]
-        if layout.device.type == "cpu":
-            bmreq_cls = TritonCPUBenchmarkRequest
-        else:
-            bmreq_cls = TritonGPUBenchmarkRequest
-        bmreq = bmreq_cls(
-            module_path=result.mod.__file__,
-            module_cache_key=result.mod.key,
-            kernel_name=f"triton_{self.name}",
-            extra_args=[*extra_args, *workspace_args, *grid],
-            num_stages=num_stages,
-            num_warps=num_warps,
-            num_consumer_groups=num_consumer_groups,
-            num_buffers_warp_spec=num_buffers_warp_spec,
-            matrix_instr_nonkdim=kwargs.get("matrix_instr_nonkdim", 0),
-            waves_per_eu=kwargs.get("waves_per_eu", 0),
-            kpack=kwargs.get("kpack", 2),
-            input_tensor_meta=TensorMeta.from_irnodes(full_input_nodes),  # type: ignore[arg-type]
-            output_tensor_meta=TensorMeta.from_irnodes(layout),
-        )
-
-        return TritonTemplateCaller(
-            kernel_hash_name,
-            full_input_nodes,
-            layout,
-            make_kernel_render,
-            result.extra.strip("-").replace("-", ", "),
-            bmreq,
-            log_info={
-                "tile_shape": str(
-                    (
-                        kwargs.get("BLOCK_M", -1),
-                        kwargs.get("BLOCK_K", -1),
-                        kwargs.get("BLOCK_N", -1),
-                    )
-                ),
-                "num_stages": num_stages,
-                "num_warps": num_warps,
-                "GROUP_M": kwargs.get("GROUP_M", -1),
-                "allow_tf32": str(kwargs.get("ALLOW_TF32", None)),
-                "acc_type": str(kwargs.get("ACC_TYPE", None)),
-                "matrix_instr_nonkdim": kwargs.get("matrix_instr_nonkdim", 0),
-                "waves_per_eu": kwargs.get("waves_per_eu", 0),
-                "kpack": kwargs.get("kpack", 2),
-            },
-            mutated_inputs=mutated_inputs,
-            workspace_arg=workspace_arg,
-            allowed_prologue_inps=result.prologue_supported_inputs,
-        )
 
 
 class ExternKernelChoice:
@@ -1751,98 +209,6 @@ class ExternKernelChoice:
         return ExternKernelCaller(
             self, input_nodes, layout, kwargs, has_out_variant=self.has_out_variant
         )
-
-
-class TritonTemplateCaller(ir.TritonTemplateCallerBase):
-    def __init__(
-        self,
-        name,
-        input_nodes,
-        layout,
-        make_kernel_render,
-        description,
-        bmreq,
-        log_info: Optional[
-            dict[str, Union[PrimitiveInfoType, list[PrimitiveInfoType]]]
-        ] = None,
-        mutated_inputs=None,
-        workspace_arg: Optional[WorkspaceArg] = None,
-        allowed_prologue_inps: Optional[OrderedSet[str]] = None,
-    ) -> None:
-        super().__init__(name, input_nodes, layout, description)
-        self.make_kernel_render = make_kernel_render
-        self.bmreq: TritonBenchmarkRequest = bmreq
-        if log_info is None:
-            log_info = {}
-        self.log_info: dict[str, Any] = log_info
-        self.log_info.update(
-            {
-                "backend": "Triton",
-                "num_stages": self.bmreq.num_stages,
-                "num_warps": self.bmreq.num_warps,
-            }
-        )
-        self.mutated_inputs = mutated_inputs
-        self.workspace_arg = workspace_arg
-        self.allowed_prologue_inps = (
-            allowed_prologue_inps if allowed_prologue_inps is not None else OrderedSet()
-        )
-
-    def benchmark(self, *args, out):
-        assert self.bmreq is not None
-        if config.profile_bandwidth_with_do_bench_using_profiling:
-            algo = self.bmreq.make_run_fn(*args, out=out)
-            return do_bench_using_profiling(algo)
-        return self.bmreq.benchmark(*args, out=out)
-
-    def precompile(self):
-        assert self.bmreq is not None
-        self.bmreq.precompile()
-
-    def __str__(self) -> str:
-        return f"TritonTemplateCaller({self.bmreq.module_path}, {self.description})"
-
-    def call_name(self):
-        return f"template_kernels.{self.name}"
-
-    def hash_key(self):
-        return "-".join(
-            [
-                self.name.rsplit("_", 1)[0],
-                self.bmreq.module_cache_key,
-            ]
-        )
-
-    def output_node(self):
-        return ir.TensorBox.create(
-            ir.TritonTemplateBuffer(
-                layout=self.layout,
-                inputs=self.input_nodes,
-                make_kernel_render=self.make_kernel_render,
-                mutated_inputs=self.mutated_inputs,
-                allowed_prologue_inps=self.allowed_prologue_inps,
-            )
-        )
-
-    def info_dict(self) -> dict[str, Union[PrimitiveInfoType, list[PrimitiveInfoType]]]:
-        """Information returned here is logged to the autotune log file when that is enabled."""
-        return self.log_info
-
-    def get_make_kernel_render(self):
-        return self.make_kernel_render
-
-    def autoheuristic_id(self):
-        type_name = "triton"
-        info = self.info_dict()
-        # TODO(AlnisM): Does tile_shape always exist?
-        tile = info["tile_shape"]
-        tile_vals = eval(tile)  # type: ignore[arg-type]
-        BLOCK_M = tile_vals[0]
-        BLOCK_K = tile_vals[1]
-        BLOCK_N = tile_vals[2]
-        num_stages = info["num_stages"]
-        num_warps = info["num_warps"]
-        return f"type={type_name}_BLOCK-M={BLOCK_M}_BLOCK-K={BLOCK_K}_BLOCK-N={BLOCK_N}_numstages={num_stages}_numwarps={num_warps}"
 
 
 class ExternKernelCaller(ChoiceCaller):
@@ -2954,7 +1320,8 @@ class AlgorithmSelectorCache(PersistentCache):
                 return {"type": "cublas", "time": timings[choice]}
 
             assert isinstance(
-                choice, torch._inductor.select_algorithm.TritonTemplateCaller
+                choice,
+                torch._inductor.codegen.triton_templates.caller.TritonTemplateCaller,
             )
 
             info = choice.info_dict()
